yarn/npm configuration files
Added the `.yarnrc` (1) and `.npmrc` (2) configuration files to enforce
project-wide standards for all core team members and contributors.
This includes the usage of the Yarn (3) and npm (4) `exact` version
resolution as well as disabling the creation of the npm
`package-lock.json` (5) file in order to prevent conflicts with the
`yarn.lock` (6) file.

References:
  (1) https://yarnpkg.com/lang/en/docs/yarnrc
  (2) https://docs.npmjs.com/files/npmrc
  (3) https://yarnpkg.com/lang/en/docs/cli/config
  (4) https://docs.npmjs.com/misc/config#save-exact
  (5) https://docs.npmjs.com/files/package-lock.json
  (6) https://yarnpkg.com/lang/en/docs/yarn-lock

Resolves GH-17