GitHub Open Source community standards
GitHub introduced a feature for recommended community standards (1).
"arctic" tries to adhere to the great Open Source Guides (2) and adapted
to the recommendations to complete the projects community profile (3).

>>> Code of Conduct

To facilitate a healthy and constructive community behavior,
"arctic" adheres and enforce a code of conduct (4).

It includes sections about

- what we believe in and how we act
- unacceptable behavior
- the responsibilities of the maintainer
- the enforcement of the Code of Conduct
- consequences for violations
- the scope of the Code of Conduct

See the GitHub Help (5) for more details about the provided
integrations.

>>> Contributing Guidelines

The [contribution guidelines (6) help to build a community that
encourages people to use, contribute to (7), and evangelize a project.

It includes sections about

- how to get started
- bug reports
- enhancement suggestions
- pull requests
- style guides
  - [JavaScript Style Guide][styleguide-js]
  - [Markdown Style Guide][styleguide-md]
  - [Git Style Guide][styleguide-git]
- credits

See the GitHub introduction blog post (6) and GitHub Help (8) for more
details about the provided integrations.

References:
  (1) https://github.com/blog/2380-new-community-tools
  (2) https://opensource.guide
  (3) https://github.com/arcticicestudio/arctic/community
  (4) https://opensource.guide/code-of-conduct
  (5) https://help.github.com/articles/adding-a-code-of-conduct-to-your-project
  (6) https://github.com/blog/1184-contributing-guidelines
  (7) https://opensource.guide/how-to-contribute
  (8) https://help.github.com/articles/setting-guidelines-for-repository-contributors

Resolves GH-21