remark-lint (#28)
Integrated `remark-lint` (1), that is built on remark (2), the powerful
Markdown processor powered by plugins like "remark-lint".

  "Ensuring the Markdown you (and contributors) write is of great
  quality will provide better rendering in all the different markdown
  parsers, and makes sure less refactoring is needed afterwards."

"remark-lint" can be used via remark-cli (3) and a rule preset.
This preset is `remark-preset-lint-arcticicestudio` (4), the custom
preset that implements the "Arctic ice Studio Markdown Style Guide" (5).

Since the custom preset is still in major version `0` note that the
version range is `>=0.x.x <1.0.0` to avoid the "SemVer Major Zero
Caveat". When defining package versions with the the carat `^` or tilde
`~` range selector it won't affect packages with a major version of `0`.
"yarn" will resolve these packages to their exact version until the
major version is greater or equal to `1`.
To avoid this caveat the more detailed version range `>=0.x.x <1.0.0`
is used to resolve all versions greater or equal to `0.x.x` but less
than `1.0.0`. This will always use the latest `0.x.x` version and
removes the need to increment the version manually on each new release.

>>> Configuration

The `.remarkrc.js` configuration file is placed in the project root as
well as the `.remarkignore` file to also define ignore pattern.

>>> Package script/task

To allow to run the Markdown linting separately a `lint:md` package
script/task has been added to be including the main `lint` script flow.

References:
  (1) https://github.com/remarkjs/remark-lint
  (2) https://remark.js.org
  (3) https://www.npmjs.com/package/remark-cli
  (4) https://github.com/arcticicestudio/remark-preset-lint-arcticicestudio
  (5) https://arcticicestudio.github.io/styleguide-markdown

Resolves GH-27