ESLint (#31)
Integrated ESLint , the "pluggable" and de-facto standard linting
utility for JavaScript.

>>> Configuration Preset

The configuration preset `@arcticicestudio/eslint-config` (2) implements
the "Arctic Ice Studio JavaScript Style" (3). It comes with the
following peer dependencies:

- `eslint` (4)
- `babel-eslint` (5)

It it built on top of `@arcticicestudio/eslint-config-base` (6) that includes various rules of the following plugins and rule presets that are therefore also required peer dependencies:

- `eslint-config-prettier` (8)
- `eslint-plugin-babel` (9)
- `eslint-plugin-import` (9)
- `eslint-plugin-jsx-a11y` (10)
- `eslint-plugin-prettier` (11)
- `eslint-plugin-react-hooks` (12)
- `eslint-plugin-react` (13)

Since "arctic" will be built with TypeScript (14),
the `@arcticicestudio/eslint-config-typescript (15) preset will be
extended to add support for "TypeScript" source file linting and
compatibility with Prettier (16) through the
`@arcticicestudio/eslint-config-typescript/prettier` extension entry
point (17).
This preset requires the following peer dependencies:

- `@typescript-eslint/eslint-plugin` (18)
- `@typescript-eslint/parser` (19)
- `typescript` (20)

Since the custom presets are still in major version `0` note that the
version range is `>=0.x.x <1.0.0` to avoid the "SemVer Major Zero
Caveat". When defining package versions with the the carat `^` or tilde
`~` range selector it won't affect packages with a major version of `0`.
"yarn" will resolve these packages to their exact version until the
major version is greater or equal to `1`.
To avoid this caveat the more detailed version range `>=0.x.x <1.0.0`
is used to resolve all versions greater or equal to `0.x.x` but less
than `1.0.0`.
This will always use the latest `0.x.x` version and removes the need to
increment the version manually on each new release.

The `.eslintrc.js` configuration file is placed in the project root
next to the `.eslintignore` file to define ignore pattern.

To allow to lint TypeScript code the `@typescript-eslint/parser` parser
is used and specified as parser (21) next to the main Babel parser (22).

To make use of the latest experimental Babel features and proposals,
`eslint-plugin-babel` (8) has be added with the following rule
configurations:

- `babel/camelcase` with level `error` - doesn't complain about optional
  chaining (`let foo = bar?.a_b;`). Note that the core rule `camelcase`
  (23) has been disabled!
- `babel/no-unused-expressions` with level `error` - doesn't fail when
  using `do` expressions or optional chaining (`a?.b()`). Note that the
  core rule `no-unused-expressions` (24) has been disabled!

See the documentation of provided rules (25) and required configurations
to use them.

>>> Webpack Import Resolving Strategy

To prepare for a better developer experience with Webpack (that will be
used later on through Gatsby) the resolvers of the
`eslint-plugin-import` (26) has been configured for the `src` and
`src/components` paths.

>>> Rule Conflicts

The `import/no-extraneous-dependencies` rule has been disabled for
Gatsby configuration files to allow to use development dependencies
since these are necessary and will be provided by either Gatsby itself
or another without being explicitly defined as package dependency.

>>> Package Scripts

To allow to lint all TypeScript and JavaScript sources separately a
`lint:ts` and `lint:js` package script has been added that will also run
with the main `lint` script flow.

To use the great "auto-fixing" (27) feature the `format:fix-ts` and
`format:fix-js` scripts have also been added.

References:
  (1) https://eslint.org
  (2) https://github.com/arcticicestudio/styleguide-javascript:packages/%40arcticicestudio/eslint-config@develop
  (3) https://github.com/arcticicestudio.github.io/styleguide-javascript
  (4) https://github.com/eslint/eslint
  (5) https://github.com/babel/babel-eslint
  (6) https://github.com/arcticicestudio/styleguide-javascript:packages/%40arcticicestudio/eslint-config-base@develop
  (7) https://github.com/prettier/eslint-config-prettier
  (8) https://github.com/babel/eslint-plugin-babel
  (9) https://github.com/benmosher/eslint-plugin-import
  (10) https://github.com/evcohen/eslint-plugin-jsx-a11y
  (11) https://github.com/prettier/eslint-plugin-prettier
  (12) https://github.com/facebook/react/tree/master/packages/eslint-plugin-react-hooks
  (13) https://github.com/yannickcr/eslint-plugin-react
  (14) https://typescriptlang.org
  (15) https://github.com/arcticicestudio/styleguide-javascript:packages/%40arcticicestudio/eslint-config-typescript@develop
  (16) https://prettier.io
  (17) https://github.com/arcticicestudio/styleguide-javascript:packages/%40arcticicestudio/eslint-config-typescript/README.md@develop#entry-points
  (18) https://github.com/typescript-eslint/typescript-eslint:packages/eslint-plugin@master
  (19) https://github.com/typescript-eslint/typescript-eslint:packages/parser@master
  (20) https://github.com/Microsoft/TypeScript
  (21) https://eslint.org/docs/user-guide/configuring#specifying-parser
  (22) https://github.com/babel/babel-eslint
  (23) https://eslint.org/docs/rules/camelcase
  (24) https://eslint.org/docs/rules/no-unused-expressions
  (25) https://github.com/babel/eslint-plugin-babel#rules
  (26) https://github.com/benmosher/eslint-plugin-import#resolvers
  (27) https://eslint.org/docs/user-guide/command-line-interface#fixing-problems

Resolves GH-30