Prettier
Integrated Prettier (1), the opinionated code formatter with support for
many languages and integrations with most editors. It ensures that all
outputted code conforms to a consistent style.

>>> Configuration

This is one of the main features of Prettier: It already provides the
best and recommended style configurations of-out-the-box.
The only option that was hanged is the print width (2). It is set to
`80` by default which not up-to-date for modern screens (might only be
relevant when working in terminals only like e.g. with Vim). It now set
to `120` used by all of Arctic Ice Studio's style guides.

The `prettier.config.js` configuration file is placed in the project
root as well as the `.prettierignore` file to also define ignore
pattern.

>>> ESLint Compatibility

To be fully compatible with ESLint, `eslint-plugin-prettier` (3) has
already been included in GH-30 as well as the set of recommended
rules (4) via the `@arcticicestudio/eslint-config/prettier` (5) and
`@arcticicestudio/eslint-config-typescript/prettier` (6) extension
entry points.

>>> Package Script

To allow to format all sources a `format:pretty` package script has been
added that'll also run in the main `format` script flow.

A new `lint:pretty` script also allows to check if all supported files
are correct formatted. It is included in the main `lint` script flow.

References:
  (1) https://prettier.io
  (2) https://prettier.io/docs/en/options.html#print-width
  (3) https://github.com/prettier/eslint-plugin-prettier
  (4) https://github.com/prettier/eslint-config-prettier
  (5) https://github.com/arcticicestudio/styleguide-javascript/blob/develop/packages/%40arcticicestudio/eslint-config/README.md#entry-points
  (6) https://github.com/arcticicestudio/styleguide-javascript/blob/develop/packages/%40arcticicestudio/eslint-config-typescript/README.md#entry-points

Resolves GH-32