"Root" core container component (data provider)
The `Root` core container is the first and one of the important main
data provider components of the main "Gatsby Theme". It represents the
base element and entry point that wraps custom application components
and serves as a data provider later on.
Some of the tasks in the future will be to

- provide context `props` and functions for the global
  styled-components (1) theme through the `ThemeProvider` (2) component
- inject global styles through "styled-components"
  `createglobalstyle` (3) function including typography (GH-5) related
  data like application-wide used fonts
- possibly data stores for state management libraries like "MobX" (4)

This initial implementation renders the passed `children` within a
`React.Fragment`.
The website package uses the component to render the "landing"
(root/index) page component.

References:
  (1) https://styled-components.com
  (2) https://www.styled-components.com/docs/api#themeprovider
  (3) https://www.styled-components.com/docs/api#createglobalstyle
  (4) https://mobx.js.org

Epics: GH-6
Resolves GH-36